pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-asr-vi"
version = "0.1.0"
description = "Vietnamese ASR finetuning for Qwen3-ASR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "qwen-asr>=0.0.6",
    "torch>=2.4",
    "transformers>=4.57",
    "peft>=0.15",
    "accelerate>=1.0",
    "datasets>=3.0",
    "huggingface-hub>=0.20",
    "wandb",
    "jiwer",
    "librosa",
    "soundfile",
    "pyyaml",
]

[project.optional-dependencies]
train = [
    "deepspeed>=0.14",
    "flash-attn>=2.5",
]
dev = [
    "pytest>=7.0",
    "ruff",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
